backend/requirements.txt
asgiref==3.7.2
boto3==1.34.86
botocore==1.34.86
certifi==2024.2.2
click==8.1.7
Django==5.0.3
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gensim==4.3.2
jmespath==1.0.1
joblib==1.3.2
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
s3transfer==0.10.1
scikit-learn==1.4.1.post1
scipy==1.10.1
six==1.16.0
smart-open==7.0.4
sqlparse==0.4.4
threadpoolctl==3.4.0
tqdm==4.66.2
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
